requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.2
coverage==5.5
cryptography==42.0.5
debugpy==1.8.1
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
et-xmlfile==1.1.0
executing==2.0.1
eyed3==0.9.7
filetype==1.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
ipykernel==6.29.4
ipython==8.22.2
ipywidgets==8.1.2
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
lxml==5.1.1
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.14.3
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
parso==0.8.3
pdf2image==1.17.0
pdfminer.six==20231228
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-pptx==0.6.23
pytz==2024.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
striprtf==0.0.26
tiktoken==0.6.0
toml==0.10.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
widgetsnbextension==4.0.10
XlsxWriter==3.2.0